% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlocus.R
\name{dlocus}
\alias{dlocus}
\title{A novel decomposition method for brain network dynamic connectivity matrices
using low-rank structure with uniform sparsity and temporal group lasso}
\usage{
dlocus(
  Y,
  q,
  V,
  n_subject,
  preprocess = TRUE,
  penalty_S = "SCAD",
  phi = 2,
  penalty_A = FALSE,
  nu = NULL,
  maxIteration = 100,
  approximation = TRUE,
  espli1 = 0.01,
  espli2 = 0.01,
  rho = 0.9,
  silent = FALSE
)
}
\arguments{
\item{Y}{Group-level connectivity data of dimension \code{N} \times \code{K}.
\code{N} is number of subjects, \code{K} is number of edges.}

\item{q}{The number of latent sources one would like to extract.}

\item{V}{The number of nodes.}

\item{n_subject}{The number of subjects in the study.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data will be preprocessed.
Defaults to \code{TRUE}.}

\item{penalty_S}{The penalty added on latent sources. Defaults to \code{"SCAD"}.}

\item{phi}{A tunning parameters with respect to the element-wise penalty on S.}

\item{penalty_A}{If \code{TRUE}, temporal smooth lasso penalty is added on loadings. Defaults to \code{FALSE}.}

\item{nu}{A numeric tunning parameter with respect to the temporal smooth lasso penalty on A. Defaults to \code{NULL}.}

\item{maxIteration}{The number of maximum iterations. Defaults to value 100.}

\item{approximation}{If \code{TRUE}, use approximated algorithm based on SVD for update.
Defaults to \code{TRUE}. Don't change if you are not sure.}

\item{espli1}{A number describing toleration for convergence on S. Defaults to value 0.01.}

\item{espli2}{A number describing toleration for convergence on A. Defaults to value 0.01.}

\item{rho}{A tuning parameter for selecting number of ranks in each subnetwork's decomposition. Defaults to value 0.9.}

\item{silent}{If \code{TRUE}, print out the penalty added on A and S. Defaults to \code{FALSE}.}
}
\value{
The resulting loading matrix \code{A} and latent sources \code{S} based on the decomposition method
}
\description{
A novel decomposition method for brain network dynamic connectivity matrices
using low-rank structure with uniform sparsity and temporal group lasso
}
