% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlocus_initial.R
\name{dlocus_initial}
\alias{dlocus_initial}
\title{Generate starting values for the iteration}
\usage{
dlocus_initial(Y, q, V, rho = 0.95, R = NULL, maxIter = 100)
}
\arguments{
\item{Y}{A demeaned concatenated group-level connectivity data after whitening if preprocess = TRUE}

\item{q}{The number of latent sources one would like to extract}

\item{V}{The number of nodes in the study}

\item{rho}{A parameter controls the similarity between initial ica results and reconstructed results}

\item{R}{A numeric vector of pre-specifying ranks of latent sources}

\item{maxIter}{The maximum iteration number when doing ica}
}
\value{
Initial values of loading matrix \code{A} and latent source \code{S}
}
\description{
Generate starting values for the iteration
}
