% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlocus_update.R
\name{dlocus_update}
\alias{dlocus_update}
\title{Node-rotation algorithm (an extremely efficient approximation version)
An extremely efficient approximation version with potentially higher performance}
\usage{
dlocus_update(
  Y,
  A,
  theta,
  q,
  K,
  V,
  n_subject,
  t_length,
  penalt_S = TRUE,
  lambda_ch,
  gamma,
  penalt_A = FALSE,
  nu,
  H_star,
  preprocess,
  silent = FALSE
)
}
\arguments{
\item{Y}{The (preprocessed) concatenated group-level connectivity data.}

\item{A}{A loading matrix from previous iteration.}

\item{theta}{A list containing ic-specific parameters and channels.}

\item{q}{The number of latent sources one would like to extract.}

\item{K}{The number of edges.}

\item{V}{The number of nodes.}

\item{n_subject}{The number of subjects in the study.}

\item{t_length}{The number of sliding windows.}

\item{penalt_S}{Penalty added on latent sources. Defaults to \code{TRUE}.}

\item{lambda_ch}{A tunning parameters with respect to the element-wise penalty on S.}

\item{gamma}{A tunning parameters with respect to the element-wise penalty on S.}

\item{penalt_A}{If \code{TRUE}, temporal smooth lasso penalty is added on loadings. Defaults to \code{FALSE}.}

\item{nu}{A tunning parameter with respect to the temporal smooth lasso penalty on A.}

\item{H_star}{dewhitening matrix.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data is preprocessed.}

\item{silent}{If \code{TRUE}, print out the penalty added on A and S. Defaults to \code{FALSE}.}
}
\value{
Updated values of loading matrix \code{A} and latent source \code{S}.
}
\description{
Node-rotation algorithm (an extremely efficient approximation version)
An extremely efficient approximation version with potentially higher performance
}
