% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psdmat_inverse}
\alias{psdmat_inverse}
\title{Helper function to calculate a perturb inverse of a matrix}
\usage{
psdmat_inverse(mat)
}
\arguments{
\item{mat}{A symmetric matrix}
}
\description{
Helper function to calculate a perturb inverse of a matrix
}
