% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SCAD_func}
\alias{SCAD_func}
\title{Helper function to calculate the scad penalty}
\usage{
SCAD_func(yvpen, lambda_ch = 0.01, gamma = 3)
}
\arguments{
\item{yvpen}{A vector upon penalization}

\item{lambda_ch}{A tunning parameter}

\item{gamm}{A tunning parameter}
}
\description{
Helper function to calculate the scad penalty
}
